+++ src/module/general/ReadFamu/src/MeshFileTransParserHmAscii.cpp
@@ -509,7 +509,7 @@ void MeshFileTransParserHmAscii::pushBack(void)
 
 int MeshFileTransParserHmAscii::getMaxNodeNo(void) const
 {
-    int retval = _xPoints.size(); // _xPoints[] is already extended to max. node number (filled with zeros where node no. is not defined)
+    int retval = (int)_xPoints.size(); // _xPoints[] is already extended to max. node number (filled with zeros where node no. is not defined)
     ASSERT0(retval >= _noOfPoints, "sorry, and internal error occured.", _outputHandler);
     return retval;
 }
